requirements.txt
flask
flask-sqlalchemy
flask-restful
Flask-Migrate
flask-smorest
marshmallow
psycopg2-binary 
python-dotenv